WIP adding helper methods, refactoring, layout and linked names in index
