+++ app/controllers/people_controller.rb
@@ -1,8 +1,5 @@
-#require 'BmiCalculationsHelper'
-
 class PeopleController < ApplicationController
-  include BmiCalculationsHelper
-  # before_action :current_user , only: [:show, :new]
+   #before_action :current_user , only: [:show, :new]
 
   # GET /people
   def index
@@ -11,15 +8,14 @@ class PeopleController < ApplicationController
 
   # GET /people/1
   def show
-    @current_measurement = @current_user.current_measurement
-    @latest_bmi = @current_measurement.bmi
+    #@current_measurement = @current_user.current_measurement
+    #@latest_bmi = @current_measurement.bmi
+
     #@current_user = Person.find(params[:id])
-    #@current_user = Person.find_by_name("Mozes Kriebel")
-    @current_measurement = @current_user.measurements.first
-    @latest_bmi = bmi_calc(@current_measurement) #if @current_user.measurements.last
-    #history should it come from elsewhere (helper?) now: in view.
-    #todo: iterate over bmi_history to calc bmi's per measurement #Q : or in view's table?
-    @bmi_history = @current_user.measurements
+    #if @current_measurement
+      #@current_user = Person.find_by_name("Mozes Kriebel")
+     # @current_measurement = current_user.measurements.first
+     # @latest_bmi = Measurement.bmi_calc(@current_measurement) #if @current_user.measurements.last
   end
 
   # GET /people/new
@@ -31,8 +27,7 @@ class PeopleController < ApplicationController
   #POST /people
   def create
     @person = Person.create(person_params)
-
-    redirect_to people_path
+    redirect_to new_person_measurement_path([@person, @measurement])
   end
 
 
@@ -42,9 +37,9 @@ class PeopleController < ApplicationController
     params.require(:person).permit(:name, :length)
   end
 
-  def current_user
-    @current_user = Person.find(params[:id])
-  end
+  # def current_user
+  #   @current_user = Person.find(params[:id])
+  # end
 
 end
