--- app/helpers/people_helper.rb
@@ -1,2 +1,0 @@
-module PeopleHelper
-end
