+++ app/views/layouts/_measurementstable.html.erb
@@ -1,7 +1,7 @@
 <div>
   <p>Your history is </p>
   <table class="table table-striped table-bordered table-condensed">
-    <% if @current_user.measurements.any? %>
+    <% if current_user.measurements.any? %>
         <thead>
         <tr>
           <th class="col-md-2">Weight</th>
@@ -9,11 +9,11 @@
           <th class="col-md-1">BMI</th>
         </tr>
         </thead>
-        <% @current_user.measurements.each do | m | %>
+        <% current_user.measurements.each do | m | %>
             <tr>
               <td><%= m.weight %> kg </td>
               <td><%= m.date.strftime("%d %b") %></td>
-              <td><%= bmi_calc(m) %></td>
+              <td><%= Measurement.bmi_calc(m) %></td>
         <% end %>
     <% end %>
   </table>
