+++ app/views/people/show.erb
@@ -1,20 +1,20 @@
 <!-- Main jumbotron for a primary marketing message or call to action -->
 <div class="jumbotron">
-    <h1>Hello <%= @current_user.name %>!</h1>
+    <h1>Hello <%= current_user.name %>!</h1>
         <div class="container">
           <div class="col-md-12">
           <h2> Your BMI is:</h2>
-            <% if @latest_bmi < 20 || @latest_bmi >25 %>
-                <h3 class="bg-danger"><%= @latest_bmi %></h3>
+            <% if Measurement.bmi_calc(current_measurement) < 20 || Measurement.bmi_calc(current_measurement) >25 %>
+                <h3 class="bg-danger"><%= Measurement.bmi_calc(current_measurement) %></h3>
             <% else %>
-                 <h3 class="bg-success"><%= @latest_bmi %></h3>
+                 <h3 class="bg-success"><%= Measurement.bmi_calc(current_measurement) %></h3>
             <% end %>
               <div class="col-md-2">
                 <p>Is this a good or a bad thing?</p>
                 <p class="bg-success">Green is a healthy bmi, </p> <p class="bg-danger"> Red is an unhealthy bmi </p>
                 </p>
               </div>
-            <%= link_to "Add your weight here", new_person_measurement_path(@current_user.id) %>
+            <%= link_to "Add your weight here", new_person_measurement_path(current_user.id) %>
           </div>
         </div>
